Improve test runner a bit
Explicitly list the toml-test files we know to fail, so it's easier to
test for regressions and the like on changes. Can also add GitHub
actions if you want.

The stdex/float6 test was already failing on master.

"make check" will run all the tests (closes #83).

Also remove the https://github.com/cktan/toml-spec-tests, based on
@iarna/toml. These are no longer maintained, and toml-lang/toml-test
contains everything that's in there.